Split oracle and UI into modular packages
