+++ OcchioOnniveggente/src/oracle/__init__.py
@@ -1,0 +1,16 @@
+from .core import *  # noqa: F401,F403
+from .core import __all__ as core_all
+from .questions import get_questions, random_question, _USED_QUESTIONS
+from .offtopic import OFF_TOPIC_RESPONSES, OFF_TOPIC_REPLIES, off_topic_reply
+
+__all__ = (
+    core_all
+    + [
+        "get_questions",
+        "random_question",
+        "_USED_QUESTIONS",
+        "OFF_TOPIC_RESPONSES",
+        "OFF_TOPIC_REPLIES",
+        "off_topic_reply",
+    ]
+)
