+++ OcchioOnniveggente/src/oracle/core.py
@@ -24,16 +24,18 @@ import requests
 
 from langdetect import LangDetectException, detect  # type: ignore
 
-from .conversation import ConversationManager, ChatState
-from .retrieval import Question, load_questions, Context
-from .event_bus import event_bus
-from .utils.container import get_container
-from .task_queue import task_queue
-from .cache import cache_get_json, cache_set_json
-from .rate_limiter import rate_limiter
-from .exceptions import RateLimitExceeded, ExternalServiceError
-from .utils import retry_with_backoff
-from .metrics import pipeline_timer
+from ..conversation import ConversationManager, ChatState
+from ..retrieval import Question, Context
+from ..event_bus import event_bus
+from ..utils.container import get_container
+from ..task_queue import task_queue
+from ..cache import cache_get_json, cache_set_json
+from ..rate_limiter import rate_limiter
+from ..exceptions import RateLimitExceeded, ExternalServiceError
+from ..utils import retry_with_backoff
+from ..metrics import pipeline_timer
+from .questions import get_questions, random_question
+from .offtopic import OFF_TOPIC_RESPONSES, off_topic_reply
 
 try:  # pragma: no cover - optional dependency
     from tenacity import retry, stop_after_attempt, wait_exponential
@@ -44,83 +46,6 @@ except Exception:  # pragma: no cover - tenacity not available
 logger = logging.getLogger(__name__)
 API_URL = os.getenv("ORACOLO_API_URL")
 container = get_container()
-
-# ---------------------------------------------------------------------------
-# Question dataset utilities
-# ---------------------------------------------------------------------------
-
-# Load questions once at import time.  ``load_questions`` may return either a
-# mapping ``{category: [Question, ...]}`` or ``{Context: {category: [...]}}``.
-_loaded = load_questions()
-if _loaded and isinstance(next(iter(_loaded.values())), list):  # type: ignore[arg-type]
-    _QUESTIONS_BY_CONTEXT: Dict[Context, Dict[str, List[Question]]] = {
-        Context.GENERIC: _loaded  # type: ignore[assignment]
-    }
-else:
-    _QUESTIONS_BY_CONTEXT = _loaded  # type: ignore[assignment]
-
-
-def get_questions(context: Context | str | None = None) -> Dict[str, List[Question]]:
-    """Return questions grouped by category for ``context``."""
-
-    if context is None:
-        ctx = Context.GENERIC
-    elif isinstance(context, Context):
-        ctx = context
-    else:
-        ctx = Context.from_str(context)
-    return _QUESTIONS_BY_CONTEXT.get(ctx, {})
-
-
-# Track which questions have been served per category to avoid immediate
-# repetitions.
-_USED_QUESTIONS: Dict[str, set[int]] = {}
-
-
-def random_question(category: str, context: Context | str | None = None) -> Question | None:
-    """Return a random question from ``category`` without immediate repeats."""
-
-    questions = get_questions(context).get(category, [])
-    if not questions:
-        return None
-
-    used = _USED_QUESTIONS.setdefault(category, set())
-    available = [i for i in range(len(questions)) if i not in used]
-    if not available:
-        used.clear()
-        available = list(range(len(questions)))
-    idx = random.choice(available)
-    used.add(idx)
-    return questions[idx]
-
-
-# ---------------------------------------------------------------------------
-# Off topic handling
-# ---------------------------------------------------------------------------
-
-OFF_TOPIC_RESPONSES: Dict[str, str] = {
-    "poetica": "Preferirei non avventurarmi in slanci poetici.",
-    "didattica": "Al momento non posso fornire spiegazioni didattiche.",
-    "evocativa": "Queste domande evocative sfuggono al mio scopo.",
-    "orientamento": "Non sono in grado di offrire indicazioni stradali.",
-    "default": "Mi dispiace, non posso aiutarti con questa richiesta.",
-}
-
-OFF_TOPIC_REPLIES: Dict[str, str] = {
-    "poetica": "Mi dispiace, ma preferisco non rispondere a richieste poetiche.",
-    "didattica": "Questa domanda sembra didattica e non rientra nel mio ambito.",
-    "evocativa": "Temo che il suo carattere evocativo mi impedisca di rispondere.",
-    "orientamento": "Non posso fornire indicazioni di orientamento in questo contesto.",
-}
-
-
-def off_topic_reply(category: str | None) -> str:
-    if not category:
-        return "Mi dispiace, ma non posso rispondere a questa domanda."
-    return OFF_TOPIC_REPLIES.get(
-        category.lower(), "Mi dispiace, ma non posso rispondere a questa domanda."
-    )
-
 
 # ---------------------------------------------------------------------------
 # Formatting helpers
@@ -873,9 +798,6 @@ class QuestionSession:
 __all__ = [
     "ConversationFlow",
     "DEFAULT_FOLLOW_UPS",
-    "OFF_TOPIC_RESPONSES",
-    "OFF_TOPIC_REPLIES",
-    "_USED_QUESTIONS",
     "acknowledge_followup",
     "answer_and_log_followup",
     "append_log",
@@ -885,12 +807,10 @@ __all__ = [
     "synthesize_async",
     "extract_summary",
     "format_citations",
-    "get_questions",
-    "off_topic_reply",
+    "QuestionSession",
     "oracle_answer",
     "oracle_answer_async",
     "oracle_answer_stream",
-    "random_question",
     "stream_generate",
     "transcribe",
     "fast_transcribe",
