+++ OcchioOnniveggente/src/oracle/offtopic.py
@@ -1,0 +1,33 @@
+from __future__ import annotations
+
+"""Utility functions and constants for handling off-topic questions."""
+
+from typing import Dict
+
+OFF_TOPIC_RESPONSES: Dict[str, str] = {
+    "poetica": "Preferirei non avventurarmi in slanci poetici.",
+    "didattica": "Al momento non posso fornire spiegazioni didattiche.",
+    "evocativa": "Queste domande evocative sfuggono al mio scopo.",
+    "orientamento": "Non sono in grado di offrire indicazioni stradali.",
+    "default": "Mi dispiace, non posso aiutarti con questa richiesta.",
+}
+
+OFF_TOPIC_REPLIES: Dict[str, str] = {
+    "poetica": "Mi dispiace, ma preferisco non rispondere a richieste poetiche.",
+    "didattica": "Questa domanda sembra didattica e non rientra nel mio ambito.",
+    "evocativa": "Temo che il suo carattere evocativo mi impedisca di rispondere.",
+    "orientamento": "Non posso fornire indicazioni di orientamento in questo contesto.",
+}
+
+
+def off_topic_reply(category: str | None) -> str:
+    """Return a canned reply for an off topic ``category``."""
+
+    if not category:
+        return "Mi dispiace, ma non posso rispondere a questa domanda."
+    return OFF_TOPIC_REPLIES.get(
+        category.lower(), "Mi dispiace, ma non posso rispondere a questa domanda."
+    )
+
+
+__all__ = ["OFF_TOPIC_RESPONSES", "OFF_TOPIC_REPLIES", "off_topic_reply"]
