+++ OcchioOnniveggente/src/ui/__init__.py
@@ -1,0 +1,6 @@
+from .core import *  # noqa: F401,F403
+from .core import __all__ as core_all
+from .realtime_ws import RealtimeWSClient
+from .utils import highlight_terms
+
+__all__ = core_all + ["RealtimeWSClient", "highlight_terms"]
